+++ src/curses_fzf/__init__.py
@@ -2,7 +2,7 @@ from .__about__ import __version__
 from .colors import Color, ColorTheme
 from .errors import CursesFzfException, CursesFzfAborted, CursesFzfAssertion, CursesFzfIndexOutOfBounds
 from .fuzzyfinder import FuzzyFinder
-from .scoring import ScoringResult, scoring_full_words
+from .scoring import ScoringResult, scoring_fzf, scoring_full_words
 
 __all__ = [
     "__version__",
@@ -14,5 +14,6 @@ __all__ = [
     "CursesFzfIndexOutOfBounds",
     "FuzzyFinder",
     "ScoringResult",
+    "scoring_fzf",
     "scoring_full_words",
 ]